test: rename banned-name test files to compliant names
