+++ tests/integration/cli/test_read_partial_tool_extra.py
