+++ tests/integration/formatters/_suite_phase_helpers_extra.py
