+++ tests/unit/cli/test_base_command_extra.py
