+++ tests/unit/core/test_query_extra.py
