+++ tests/unit/core/test_query_library_extra.py
