+++ tests/unit/core/test_query_loader_supplement_extra.py
