+++ tests/unit/formatters/test_go_formatter_extra.py
