+++ tests/unit/formatters/test_ruby_formatter_extra.py
