+++ tests/unit/formatters/test_sql_formatter_extra.py
