+++ tests/unit/languages/test_javascript_plugin_extra.py
