+++ tests/unit/languages/test_typescript_extractor_extra.py
