+++ tests/unit/mcp/test_error_recovery_extra.py
