+++ tests/unit/mcp/test_gitignore_detector_extra.py
