+++ tests/unit/mcp/test_server_comprehensive.py
@@ -1,12 +1,12 @@
 """Re-export aggregator for split test modules."""
 
-from test_server_comprehensive_init import (  # noqa: F401
+from test_server_init_extra import (  # noqa: F401
     TestTreeSitterAnalyzerMCPServerCodeAnalysis,
     TestTreeSitterAnalyzerMCPServerCreation,
     TestTreeSitterAnalyzerMCPServerFileMetrics,
     TestTreeSitterAnalyzerMCPServerInitialization,
 )
-from test_server_comprehensive_tools import (  # noqa: F401
+from test_server_tools_extra import (  # noqa: F401
     TestMCPServerUtilities,
     TestTreeSitterAnalyzerMCPServerProjectPath,
     TestTreeSitterAnalyzerMCPServerRuntime,
