+++ tests/unit/mcp/test_server_init_extra.py
