+++ tests/unit/mcp/test_server_tools_extra.py
