+++ tests/unit/mcp/tools/test_element_extractor_extra.py
