+++ tests/unit/utils/test_encoding_utils_extra.py
